common_utils and tests
