+++ services/maker/maker/AssetData.py
@@ -1,9 +1,9 @@
 import os
 import json
+
 ASSET_CONFIG_FILE = os.environ.get('ASSET_CONFIG_FILE', 'asset_config.json')
 
 class AssetData():
-
 
     address_to_ticker = {}
 
@@ -13,5 +13,5 @@ class AssetData():
             for asset_data in asset_config_f['assets']:
                 self.address_to_ticker[asset_data['address']] = asset_data['symbol']
 
-    def get_ticker_with_address(self, address):
+    def get_ticker_with_address(self, address: str) -> str:
         return self.address_to_ticker[address]
